requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
colorama==0.4.4
decorator==5.0.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-ckeditor==6.1.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-multiselectfield==0.1.12
gunicorn==20.1.0
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycodestyle==2.7.0
Pygments==2.10.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.3.0
